POK-59 Updated folder structure for iOS and implement storyboards for login and home navigation
