+++ PokeAppiOS/Controllers/LoginViewController.cs
@@ -1,25 +1,22 @@
-﻿using CoreGraphics;
-using Foundation;
-using SharedCode.Services;
+﻿using System;
 using SharedCode.Model;
-using System;
-using UIKit;
+using SharedCode.Services;
 using SharedCode.Util;
-
-namespace PokeAppiOS
+using UIKit;
+namespace PokeAppiOS.Controllers
 {
-    public partial class ViewController : UIViewController
-    {
-        static LoginService loginService = new LoginService();
+	public partial class LoginViewController : UIViewController
+	{
+        LoginService loginService = LoginService.shared;
 
 
-        public ViewController (IntPtr handle) : base (handle)
+        public LoginViewController(IntPtr handle) : base(handle)
         {
         }
 
-        public override void ViewDidLoad ()
+        public override void ViewDidLoad()
         {
-            base.ViewDidLoad ();
+            base.ViewDidLoad();
             Title = "Login View";
 
             LoginButton.TouchUpInside += LoginButton_TouchUpInside;
@@ -47,9 +44,9 @@ namespace PokeAppiOS
             }
         }
 
-        public override void DidReceiveMemoryWarning ()
+        public override void DidReceiveMemoryWarning()
         {
-            base.DidReceiveMemoryWarning ();
+            base.DidReceiveMemoryWarning();
             // Release any cached data, images, etc that aren't in use.
         }
     }
