+++ PokeAppiOS/SceneDelegate.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Foundation;
 using PokeAppiOS;
+using PokeAppiOS.Controllers;
 using UIKit;
 
 namespace PokeAppiOS {
@@ -76,17 +77,14 @@ namespace PokeAppiOS {
 
 		public void SegueToHome()
 		{
-			// For demo purposes initializes 'SecondViewController' as UIViewControler
-			// but could be replaced by storyboard
-            UIViewController viewController = new SecondViewController();
-            UINavigationController navigationMain = new UINavigationController(viewController);
-            Window.RootViewController = navigationMain;
-            Window.MakeKeyAndVisible();
+            var mainStoryboard = UIStoryboard.FromName("Home", null).InstantiateInitialViewController() as UIViewController;
+            Window.RootViewController = mainStoryboard;
+			Window.MakeKeyAndVisible();
         }
 
 		private void SegueToLogin()
 		{
-			var mainStoryboard = UIStoryboard.FromName("Main", null).InstantiateInitialViewController() as UIViewController;
+            var mainStoryboard = UIStoryboard.FromName("Login", null).InstantiateInitialViewController() as UIViewController;
             Window.RootViewController = mainStoryboard;
             Window.MakeKeyAndVisible();
         }
