--- PokeAppiOS/SecondViewController.xib
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="SecondViewController">
-            <connections>
-                <outlet property="LabelTest" destination="Gtg-R0-ffG" id="FC1-Ej-dgo"/>
-                <outlet property="LogoutButton" destination="yem-QQ-J9n" id="CIr-nh-2HL"/>
-                <outlet property="view" destination="2" id="RRd-Eg-VrN"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="2">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <subviews>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Replace this view by the pokemon list" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Gtg-R0-ffG">
-                    <rect key="frame" x="54.333333333333343" y="416" width="284.66666666666663" height="20.333333333333314"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="yem-QQ-J9n">
-                    <rect key="frame" x="120" y="451.33333333333331" width="153" height="35"/>
-                    <state key="normal" title="Button"/>
-                    <buttonConfiguration key="configuration" style="filled" title="Logout"/>
-                </button>
-            </subviews>
-            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="yem-QQ-J9n" firstAttribute="leading" secondItem="2" secondAttribute="leading" constant="120" id="Jnz-M2-6eV"/>
-                <constraint firstItem="yem-QQ-J9n" firstAttribute="top" secondItem="Gtg-R0-ffG" secondAttribute="bottom" constant="15" id="NLU-jt-TD3"/>
-                <constraint firstItem="Gtg-R0-ffG" firstAttribute="centerY" secondItem="2" secondAttribute="centerY" id="QeX-eK-tqq"/>
-                <constraint firstItem="Gtg-R0-ffG" firstAttribute="centerX" secondItem="2" secondAttribute="centerX" id="VCy-ys-kls"/>
-                <constraint firstAttribute="trailing" secondItem="yem-QQ-J9n" secondAttribute="trailing" constant="120" id="djc-6a-lcH"/>
-            </constraints>
-            <point key="canvasLocation" x="-54" y="-53"/>
-        </view>
-    </objects>
-</document>
