--- PokeAppiOS/View/LoginViewController.cs
@@ -1,27 +1,0 @@
-﻿using System;
-
-using UIKit;
-
-namespace PokeAppiOS.View
-{
-	public partial class LoginViewController : UIViewController
-	{
-		public LoginViewController () : base ("LoginViewController", null)
-		{
-		}
-
-		public override void ViewDidLoad ()
-		{
-			base.ViewDidLoad ();
-			// Perform any additional setup after loading the view, typically from a nib.
-		}
-
-		public override void DidReceiveMemoryWarning ()
-		{
-			base.DidReceiveMemoryWarning ();
-			// Release any cached data, images, etc that aren't in use.
-		}
-	}
-}
-
-
